Break mapping out into its own module
